Bug: Port is ignored when specified from extra_args (#84)
This PR addresses #80 partially. If the user specifies the connection port as a command-line option via extra_args, it is assumed to be an intentional choice, and this option takes precedence. If the port is specified both through the port parameter and extra_args, the port from extra_args is chosen, but a warning is displayed.

This PR also includes clarifications in the Open Connection documentation about the correct use of hyphens in command-line options. The examples previously included two hyphens (--), but the correct usage is one hyphen (-).

Finally, an internal redesign of py3270.py has made the module more concise.